requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
openai-whisper==20231117
faiss-cpu==1.8.0
sentence-transformers==3.0.1
groq>=0.9.0
langdetect==1.0.9
librosa==0.10.2
numpy==1.26.4
pydantic==2.7.1
